// src/components/llm-components/types.ts
export type ChartType = "bar" | "line" | "area" | "pie";
// ...
export interface ChartYKey {
  key?: string;
  label?: string;
  color?: string;
}

export interface ChartConfig {
  xKey?: string;
  yKeys?: ChartYKey[];
}

export interface ChartComponent {
  type: "chart";
  chartType?: ChartType;
  title?: string;
  description?: string;
  data?: Record<string, string | number>[];
  config?: ChartConfig;
}